moving helpers to component
