+++ samples/23.react-live-canvas/src/utils/useLiveCanvas.jsx
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { InkingManager, InkingTool } from "@microsoft/live-share-canvas";
+import { InkingManager } from "@microsoft/live-share-canvas";
 import { useCallback, useEffect, useState } from "react";
 
 /**
@@ -15,46 +15,6 @@ import { useCallback, useEffect, useState } from "react";
 export const useLiveCanvas = (liveCanvas, hostingElement) => {
   const [inkingManager, setInkingManager] = useState(undefined);
   const [error, setError] = useState(undefined);
-
-  const setToPen = useCallback(() => {
-    inkingManager.tool = InkingTool.pen
-  }, [inkingManager])
-
-  const setToLaserPointer = useCallback(() => {
-    inkingManager.tool = InkingTool.laserPointer
-  }, [inkingManager])
-
-  const setToHighlighter = useCallback(() => {
-    inkingManager.tool = InkingTool.highlighter
-  }, [inkingManager])
-
-  const setToEraser = useCallback(() => {
-    inkingManager.tool = InkingTool.pointEraser
-  }, [inkingManager])  
-
-  const setToBlackBrush = useCallback(() => {
-    inkingManager.penBrush.color = { r: 0, g: 0, b: 0 }
-  }, [inkingManager])
-
-  const setToYellowBrush = useCallback(() => {
-    inkingManager.penBrush.color = { r: 255, g: 252, b: 0 }
-  }, [inkingManager])
-
-  const setToBlueBrush = useCallback(() => {
-    inkingManager.penBrush.color = { r: 0, g: 0, b: 255, a: 1 }
-  }, [inkingManager]) 
-
-  const setToGreenBrush = useCallback(() => {
-    inkingManager.penBrush.color = { r: 0, g: 255, b: 0 }
-  }, [inkingManager])
-
-  const setToRedBrush = useCallback(() => {
-    inkingManager.penBrush.color = { r: 255, g: 0, b: 0 }
-  }, [inkingManager])
-  
-  const clearCanvas = useCallback(() => {
-    inkingManager.clear();
-  }, [inkingManager])
 
   const startInkingManager = useCallback(async () => {
     if (!liveCanvas || !hostingElement) {
@@ -82,16 +42,6 @@ export const useLiveCanvas = (liveCanvas, hostingElement) => {
   return {
     canvasReady: !!inkingManager,
     inkingManager,
-    setToPen,
-    setToLaserPointer,
-    setToHighlighter,
-    setToBlackBrush,
-    setToBlueBrush,
-    setToYellowBrush,
-    setToGreenBrush,
-    setToRedBrush,
-    setToEraser,
-    clearCanvas,
     error,    
   }
 }
